+++ packages/spor-theme-react-native/src/components/button.tsx
@@ -37,6 +37,7 @@ export const buttonVariants = {
     backgroundColor: "transparent",
     color: "darkGrey",
     fontWeight: "normal",
+    px: 0,
   },
 };