+++ packages/spor-accordion-react-native/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vygruppen/spor-accordion-react-native
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [8dbf17e]
+  - @vygruppen/spor-theme-react-native@0.2.2
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [bfb0d80]
+  - @vygruppen/spor-theme-react-native@0.2.1
+
 ## 0.0.3
 
 ### Patch Changes