+++ packages/spor-badge-react-native/CHANGELOG.md
@@ -1,6 +1,21 @@
 # @vygruppen/spor-badge-react-native
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [8dbf17e]
+  - @vygruppen/spor-theme-react-native@0.2.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [bfb0d80]
+  - @vygruppen/spor-theme-react-native@0.2.1
+
 ## 1.0.0
+
 ### Major Changes
 
 - c54d418: Lagt til en ny pakke for Badges