+++ packages/spor-button-react-native/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @vygruppen/spor-button-react-native
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [8dbf17e]
+  - @vygruppen/spor-loader-react-native@0.1.5
+  - @vygruppen/spor-theme-react-native@0.2.2
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [bfb0d80]
+  - @vygruppen/spor-theme-react-native@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes