+++ packages/spor-card-react-native/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vygruppen/spor-card-react-native
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [8dbf17e]
+  - @vygruppen/spor-theme-react-native@0.2.2
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [bfb0d80]
+  - @vygruppen/spor-theme-react-native@0.2.1
+
 ## 0.0.6
 
 ### Patch Changes